+++ sparkles-core/src/headers.rs
@@ -1,6 +1,7 @@
 use alloc::string::{String, ToString};
 use serde::{Deserialize, Serialize};
 use crate::local_storage::id_mapping::IdStoreMap;
+use crate::{Timestamp, TimestampProvider};
 
 /// This header describe byte buffer, full of encoded sparkles events.
 /// This header is thread-local, each thread has its own header and buffer.
@@ -24,18 +25,29 @@ pub struct ThreadInfo {
 #[derive(Serialize, Deserialize, Clone, Debug)]
 pub struct SparklesEncoderInfo {
     pub ver: u32,
-    pub counts_per_ns: f64,
     pub process_name: String,
     pub pid: u32,
+    pub timestamp_max_value: u64
+}
+
+impl SparklesEncoderInfo {
+    pub fn new(process_name: String, pid: u32) -> Self {
+        Self {
+            pid,
+            process_name,
+            ver: crate::consts::ENCODER_VERSION,
+            timestamp_max_value: Timestamp::MAX_VALUE,
+        }
+    }
 }
 
 impl Default for SparklesEncoderInfo {
     fn default() -> Self {
         Self {
-            ver: crate::consts::ENCODER_VERSION,
-            counts_per_ns: 2.495,
             process_name: "unknown".to_string(),
             pid: 0,
+            ver: crate::consts::ENCODER_VERSION,
+            timestamp_max_value: Timestamp::MAX_VALUE,
         }
     }
 }