+++ sparkles-parser/src/lib.rs
@@ -1,16 +1,15 @@
 mod perfetto_format;
 mod consts;
+mod decoder;
 
 use std::cmp::min;
 use std::collections::BTreeMap;
 use std::io::{Read, Write};
-use std::mem;
 use log::{debug, error, info, warn};
-use ringbuf::storage::Heap;
-use ringbuf::traits::{Consumer, Observer, Producer};
 use thiserror::Error;
 use sparkles_core::headers::{LocalPacketHeader, SparklesEncoderInfo};
 use sparkles_core::local_storage::id_mapping::EventType;
+use crate::decoder::StreamFrameDecoder;
 use crate::ParseError::Decode;
 use crate::perfetto_format::PerfettoTraceFile;
 
@@ -22,6 +21,7 @@ pub struct SparklesParser {
     total_transport_bytes: u64,
 
     encoder_info: Option<SparklesEncoderInfo>,
+    ticks_per_ns: Option<f64>,
 
     event_parsers: BTreeMap<u64, ThreadParserState>,
 }
@@ -83,7 +83,10 @@ impl SparklesParser {
         info!("Begin parsing... Encoder info: {:?}", encoder_info);
 
         let mut trace_res_file = PerfettoTraceFile::new(encoder_info.process_name, encoder_info.pid);
-        let counts_per_ns = encoder_info.counts_per_ns;
+        let ticks_per_ns = self.ticks_per_ns.unwrap_or_else( || {
+            warn!("Did not find timestamp frequency in decoded stream! Using default values");
+            1.0
+        });
         // iterate over all threads
         for (&thread_ord_id, parser_state) in &mut self.event_parsers {
             let thread_name = parser_state.thread_name.clone().unwrap_or("".to_string());
@@ -107,7 +110,7 @@ impl SparklesParser {
                         parser_state.cur_tm += dif_tm;
                     }
                     // add to trace file
-                    let timestamp = (parser_state.cur_tm as f64 / encoder_info.counts_per_ns) as u64;
+                    let timestamp = (parser_state.cur_tm as f64 / ticks_per_ns) as u64;
                     match event {
                         TracingEvent::Instant(id, _) => {
                             let (ev_name, _) = &header.id_store.tags[*id as usize];
@@ -149,12 +152,12 @@ impl SparklesParser {
             }
         }
 
-        let events_per_sec = total_events as f64 / ((max_timestamp - min_timestamp) as f64 / counts_per_ns) * 1_000_000_000.0;
-        let events_per_sec_covered = total_events as f64 / (covered_dur as f64 / counts_per_ns) * 1_000_000_000.0;
+        let events_per_sec = total_events as f64 / ((max_timestamp - min_timestamp) as f64 / ticks_per_ns) * 1_000_000_000.0;
+        let events_per_sec_covered = total_events as f64 / (covered_dur as f64 / ticks_per_ns) * 1_000_000_000.0;
         info!("Total events: {}", total_events);
         info!("Events per second (global): {} eps", events_per_sec);
         info!("Events per second (covered): {} eps", events_per_sec_covered);
-        info!("Average event duration: {} ns", covered_dur as f64 / counts_per_ns / total_events as f64);
+        info!("Average event duration: {} ns", covered_dur as f64 / ticks_per_ns / total_events as f64);
         info!("Average bytes per event: {} bytes", self.total_event_bytes as f64 / total_events as f64);
         info!("Average transport bytes per event: {} bytes", self.total_transport_bytes as f64 / total_events as f64);
 
@@ -267,6 +270,15 @@ impl SparklesParser {
                     self.thread_parser_state(thread_ord_id).missed_events.push((start, dur));
 
                 },
+                0x03 => {
+                    let mut bytes = [0u8; 8];
+                    con.read_exact(&mut bytes)?;
+                    let ticks_per_sec = u64::from_le_bytes(bytes);
+                    let ticks_per_ns = ticks_per_sec as f64 / 1_000_000_000.0;
+                    info!("Got timestamp frequency: {:?} t/ns", ticks_per_ns);
+
+                    self.ticks_per_ns = Some(ticks_per_ns);
+                }
                 0xff => {
                     info!("Client was gracefully disconnected!");
 
@@ -290,124 +302,4 @@ pub enum TracingEvent {
     Instant(TracingEventId, u64),
     RangePart(TracingEventId, u64, u8),
     UnnamedRangeEnd(u64, u8)
-}
-
-pub struct StreamFrameDecoder {
-    state: ParsingState,
-    buf: ringbuf::LocalRb<Heap<u8>>
-}
-
-impl Default for StreamFrameDecoder {
-    fn default() -> Self {
-        Self {
-            state: ParsingState::NewFrame,
-            buf: ringbuf::LocalRb::new(PARSER_BUF_SIZE)
-        }
-    }
-}
-
-#[derive(Copy,Clone, Default, Debug)]
-#[derive(PartialEq)]
-pub enum ParsingState {
-    #[default]
-    NewFrame,
-    DifTmLen(TracingEventId),
-
-    /// id, dif_tm_len
-    DifTm(TracingEventId, usize),
-
-    RangeOrdId(Option<TracingEventId>, usize),
-    RangeTm(Option<TracingEventId>, usize, u8)
-}
-
-impl StreamFrameDecoder {
-    pub fn try_decode_event(&mut self) -> Result<TracingEvent, bool> {
-        let available_bytes_len = self.buf.occupied_len();
-
-        let (ev, new_state) = match mem::take(&mut self.state) {
-            ParsingState::NewFrame if available_bytes_len >= 1 => {
-                let ev_id = self.buf.try_pop().unwrap();
-                (None, ParsingState::DifTmLen(ev_id))
-            }
-            ParsingState::DifTmLen(ev) if available_bytes_len >= 1 => {
-                let dif_tm_len = self.buf.try_pop().unwrap();
-
-                let is_range_event = dif_tm_len & 0b1000_0000 != 0;
-                let is_unnamed_range_end = dif_tm_len & 0b0100_0000 != 0;
-                let dif_tm_len = (dif_tm_len & 0b0000_1111) as usize;
-
-                if is_range_event {
-                    if is_unnamed_range_end {
-                        (None, ParsingState::RangeOrdId(None, dif_tm_len))
-                    }
-                    else {
-                        (None, ParsingState::RangeOrdId(Some(ev), dif_tm_len))
-                    }
-                }
-                else {
-                    (None, ParsingState::DifTm(ev, dif_tm_len))
-                }
-            }
-            ParsingState::DifTm(ev, dif_tm_len) if available_bytes_len >= dif_tm_len => {
-                let mut buf = [0u8; 8];
-                self.buf.pop_slice(&mut buf[..dif_tm_len]);
-                let dif_tm = u64::from_le_bytes(buf);
-                (Some(TracingEvent::Instant(ev, dif_tm)), ParsingState::NewFrame)
-            }
-            ParsingState::RangeOrdId(ev, dif_tm_len) if available_bytes_len >= 1 => {
-                let ord_id = self.buf.try_pop().unwrap();
-
-                (None, ParsingState::RangeTm(ev, dif_tm_len, ord_id))
-            }
-            ParsingState::RangeTm(ev_id, dif_tm_len, ord_id) if available_bytes_len >= dif_tm_len => {
-                let mut buf = [0u8; 8];
-                self.buf.pop_slice(&mut buf[..dif_tm_len]);
-                let dif_tm = u64::from_le_bytes(buf);
-
-                let ev = if let Some(id) = ev_id {
-                    Some(TracingEvent::RangePart(id, dif_tm, ord_id))
-                }
-                else {
-                    Some(TracingEvent::UnnamedRangeEnd(dif_tm, ord_id))
-                };
-                (ev, ParsingState::NewFrame)
-            }
-            state => {
-                // Not enough bytes
-                self.state = state;
-                return Err(true)
-            }
-        };
-
-        self.state = new_state;
-        if let Some(ev) = ev {
-            Ok(ev)
-        }
-        else {
-            Err(false)
-        }
-    }
-
-    pub fn decode_many(&mut self, bytes: &[u8]) -> Vec<TracingEvent> {
-        self.buf.push_slice(bytes);
-        let mut events = Vec::new();
-        // Try parse as many events as possible
-        loop {
-            match self.try_decode_event() {
-                Ok(ev) => {
-                    events.push(ev);
-                }
-                Err(finish) => {
-                    if finish {
-                        return events;
-                    }
-                }
-            }
-        }
-    }
-
-    pub fn ensure_buf_end(&mut self) {
-        assert!(self.buf.is_empty());
-        assert_eq!(self.state, ParsingState::NewFrame);
-    }
 }